requirements.txt
Flask==2.3.3
APScheduler==3.10.4
bcrypt==4.1.3
PyJWT==2.8.0